Refactor A1CommandBase, improve device type handling
Moved A1CommandBase to the a1 package and updated imports accordingly. Enhanced ConnectionManager to handle 0xA0 responses based on device type (ac or a1). Added HIGH4 fan speed to ACStringCommands. Improved power state handling in A1CommandSet and added related test. Updated status handling in device handlers and adjusted tests for new logic.
