+++ bundles/org.openhab.binding.mideaac/src/main/java/org/openhab/binding/mideaac/internal/handler/MideaDehumidifierHandler.java
@@ -105,6 +105,8 @@ public class MideaDehumidifierHandler extends AbstractMideaHandler implements Hu
     protected void refreshDeviceState() {
         try {
             connectionManager.getStatus(this);
+            // If we reach here, the device is online.
+            updateStatus(ThingStatus.ONLINE);
         } catch (MideaAuthenticationException e) {
             updateStatus(ThingStatus.OFFLINE, ThingStatusDetail.CONFIGURATION_ERROR, e.getMessage());
         } catch (MideaConnectionException e) {
